Phase4/backend/pyproject.toml
[tool.poetry]
name = "todo-ai-backend"
version = "0.1.0"
description = "Backend for Todo AI Chatbot"
authors = ["Developer <developer@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlmodel = "^0.0.16"
openai = "^1.3.5"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
alembic = "^1.13.1"
neon = "^0.3.0"
asyncpg = "^0.29.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
sse-starlette = "^1.6.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
httpx = "^0.25.2"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "scripts.dev:main"